실제 거리 문제 해결 , 운동 씬 추가 정리 조정 중 +

+++ gymchairaos-unity-develop/Assets/TreadmillProject/ZBWorks/Scripts/Balance/DistanceCount.cs
@@ -10,8 +10,8 @@ namespace ZB.Balance
         [SerializeField] WheelChairGuage m_wheelChairGuage;
         [SerializeField] TextMeshProUGUI m_tmp_distanceShow;
 
-        public float m_NowDistance { get => m_nowDistance; }
-        [SerializeField] private float m_nowDistance;
+        public double m_NowDistance { get => m_nowDistance; }
+        [SerializeField] private double m_nowDistance;
 
         [ContextMenu("Start")]
         public void CountStart()
